Rename service/ to input_transform/
"service" said nothing about what the package does -- every folder in src/ is a
service in some sense. This one is specifically the inbound half: read from the
hospital, normalise to the dict shapes the agents consume.

Named input_transform rather than fhir_input_transform because only 2 of its 10
modules read FHIR. clinical.py and lab.py do; financial, ot, ambulance,
appointments and pharmacy read plain REST with no FHIR involved (two of them say so
in their own docstrings), and staff/ventilator pass raw table rows straight through.
Labelling the whole package FHIR would have misled on 7 files. The package
docstring records that reasoning so the name doesn't drift back.

Mechanical: git mv so history follows, 20 import sites, and the prose references in
17 files plus the logging_config section comment.

Verified from a clean venv: 75 tests pass, 88 routes, 85/85 OpenAPI summaries.
